requirements.txt
vkbottle>=4.3.12
python-dotenv>=1.0.0
requests>=2.31.0
